User-agent changes
This change allows the user agent to be longer than 255 chars (there are legitimate
user agents that can be longer than that) but caps it with a validation at 2048 (which
matches the `feedback` app).

It looks like this worked before because the `support` app truncated any overrunning
user agents automatically.