Add worker for uploading to the Performance Platform
There's one job per worker, per day. This is so that if the dataset for
a particular service doesn't exist yet (as the service is new), then
the metrics for the other services upload without a problem, and the upload
for the new service is retried repeatedly (by sidekiq) until it succeeds or reaches
the maximum number of retries.