Add rake task for running frequent deduping
There is a nightly task to dedupe anonymous feedback from the previous
day. However, there are still duplicates that show up when feedback
for the current day is viewed (this is especially obvious for user
satisfaction feedback).

This change enables a deduplication task to run every 5 minutes which eliminates
the most recent duplicates.