Deduplication bugfix: add all anonymous feedback types
There are several types of anonymous feedback (problem reports,
service feedback, long-form contacts) and they are all persisted in the
`anonymous_contacts` table using STI.

The deduplication process only references the `AnonymousContact` superclass,
but in order to mark a record as a duplicate, Rails needs to instanciate it
and if the class specified in the `type` column of the record isn't found
within the app, an exception is thrown.

This means that for de-duplication to actually work properly, the classes
for all types of feedback need to be present and loaded.