Make the "path" column a VARCHAR2(2048)
A path (the location of where a piece of anonymous feedback has been left)
is currently stored as an ActiveRecord `text` field, which maps to a MySQL `mediumtext`
field. In order to be able to put an index on this field, it needs to turn into a `varchar`.

The model validation on that field is a maximum length of 2048, and in the production instance,
the maximum length is 2030 chars. This makes it safe to change the column to be a `varchar2(2048)`.