Allow longer referrers
Before this change, `referrer` information for anonymous feedback
was being saved to a `varchar(255)` field. The rails 3.2 version of
the activerecord MySQL adapter truncated the referrer info, whereas
the rails 4.1 version raises an exception.

The `referrer` validation ensures that the field isn't longer than 2048 chars,
and the longest value in the production database is 2030 chars, so this limit should
be safe.