Introduce the concept of an enhanced Content API
This class tries to paper over the fact that GOV.UK doesn't yet have
a consistent content API for a large portion of the published content.

The class has two main responsibilities:

- for a given GOV.UK path, work out the path that has an entry in an API
  eg for `/contact-ukvi/overview`, only `/contact-ukvi` has an API entry

- for a given GOV.UK path, fetch the list of organisations associated with it.

There are 5 separate handlers for different formats:

- mainstream content
- organisations
- worldwide organsiations
- Depts & Policy content
- GDS-owned content (search, browse, user support pages, etc)

This isn't designed to be a 100% solution, but aimed at identifying the
relevant organisations for vast majority of content on GOV.UK.

The point of doing this is to be able to work out which organisation may
be interested in a particular piece of feedback (that's associated with a GOV.UK
path).