Problem report daily totals calculated in UTC days
Before this fix, the reporting interval for problem report daily
totals was dependent on the day and timezone of the running system.
This meant that the tests failed on an OS X system running in BST.

After this change, the reporting interval is in UTC, and the timestamp
format is explicitly defined to remove system-dependent behaviour.