Migrate the corporate content PP upload from Support app
The monthly upload of corporate content stats to the PP used to live
and run within the Support app, but belongs within the Support API,
closer to the data.

This data supports provides data to the feedback modules of the
performance platform GOV.UK web traffic dashboards
(https://www.gov.uk/performance/web-traffic).