Inline the default scope on AnonymousContact
Not having the default scope means that ActiveRecord lookups are more explicit
about what data is returned.

The `where.not(path: nil)` condition is no longer necessary because
path is not-null at the DB level.
The `only_actionable` scope has been added in all the relevant places, namely
controllers and metrics rollups.