Remove feedback records with invalid dates
There was a short scan of GOV.UK in November 2012 using the Vega tool.
This populated some feedback forms various maliciously crafted strings
and submitted them. There doesn’t seem to be any side effects from
this, except that somehow created_at and updated_at dates were set to
blank for just 3 of the records.

The information provided is meaningless, so this migration simply
deletes the records.

It will find other records with blank (not NULL) dates and delete them,
but as a safeguard it’ll just refuse to do anything if it finds more
than 10 records.

This work is needed because attempting to import the data into Postgres
in this state will make it fail the NOT NULL constraints on these
columns.