Add new index for created_at and path
This uses a Postgres-specific index definition, so we need to execute
SQL directly. AR’s SchemaDumper cannot reverse engineer this into a
`schema.rb` file, so for Postgres we need to use the sql schema format.

Added a convenience method on `SupportApi` to determine what mode we’re
running under. Once this porting work has been completed and we remove
Mysql support, this can be removed entirely.