MySQL to PostgreSQL synchronisation script
A rake task to copy all data incrementally from MySQL to PostgreSQL.

Relies on all tables being append-only, which is the case here.

Uses the Sequel gem as it’s far better and easier than trying to
persuade the AR internal API to convert booleans etc between the two
formats.

Requires both connections to be configured as
`{mysql,postgresql}_#{Rails.env}` in database.yml

The dependencies on sequel and ruby-progressbar (and the script itself)
can be removed once this transition is complete.

The script can be interrupted and resumed when needed, it’ll resume
from the last id inserted into Postgres.