Make pagination more performant
For queries that will return a lot of results, fetching the initial
result count is very expensive (20s for over 3 million results).

Luckily, Rails can do a `COUNT` with `LIMIT` in PostgreSQL. If you
specify a limit as part of a scope then call `.count`, it will generate
SQL like this:

```
SELECT COUNT(count_column) FROM (SELECT 1 AS count_column FROM
"anonymous_contacts" WHERE path LIKE '/done/vehicle-tax%' LIMIT 10000)
subquery_for_count;
```

This returns in around 20ms on a dev VM for a query that would return 3
million rows if unlimited. It provides 10000 responses over 200 pages
for the user, with the last page rendering in about 100ms.

If the user requires all responses, they can still export them all as a
CSV. Currently only 28 paths have more than 10000 responses, all of
which are Done pages which need to be refactored at some point to
display only those with comments.

This should be coupled with frontend changes, providing copy informing
the user to download a CSV if there's more than 10000 responses.