Bugfix: ignore errors from a dataset not existing in PP
As of Dec 2015, creating service feedback datasets for each new transaction
in the Performance Platform is a manual process done by the PP team.
It's often the case that feedback starts coming into the Support API for
transactions that don't yet have a dataset for collecting this data.

Until the dataset is created, the Support API should silently consume the
failure. The data is still retained in the Support API DB, in case it needs
to be backfilled to the PP.