Allow export of Global statistics
This generates and emails (as an attachment) a CSV file containing two
fields; `date` and `report_count`, for all anonymous feedback submitted
to GOV.UK.

The aim here is to provide a mechanism to export this data so that
statistical analysis can be performed. At the moment this analysis
happens on data from zendesk; we'd like to avoid sending all types of
anonymous feedback to zendesk, so we need to have another way to provide
the data.