Group organisation report summaries by path
In the database for the `support-api`, it's possible for there to be
multiple `ContentItem`'s for a given `path`. Case in point:

    irb(main):005:0> ContentItem.where(path: '/calculate-your-holiday-entitlement').count
    => 255

When a problem report is raised, a representation of the content item
relating to that problem report is persisted to the database.

The process of accessing this content item is pretty convoluted, and
reading the code it would *appear* to be trying to update the existing
content item in the database if one already exists, but it appears that
this isn't happening.

Unpicking this code requires a significant refactor and some reasonable
understanding of the GOV.UK content API/content stores/publishing API,
which I don't have yet, but in the meantime let's fix the user facing
bug.