Refactor generating ServiceFeedback aggregates
These aggregations are reported to the performance platform. Previously,
the logic for "scoping" the result set and "presenting" this data as a
hash was mixed into both the model
(in ServiceFeedback#aggregates_by_rating) and the aggregation class
(ServiceFeedbackAggregatedMetrics#aggregates).

Instead, let's have the class return a scoped recordset, and the
injection of null values and transformation into a structured hash
happen solely in the aggregation class.