Refactor test for ServiceFeedbackAggregator
First, this class lives in `lib/` not `models/` so let's move it to that
location in the test suite.

In general, putting the execution of the operation outside the test
block (per the previous spec, putting the `aggregator.run` statement in
the `before do` block), makes the action under test more opaque.

It also makes it more difficult to resolve the order in which other,
competing operations complete before the `aggregator.run` operation. It
might, for example, trigger after one context before block, but before
a second.

Therefore, move the `aggregator.run` command inside the spec blocks.

Rethinking the contexts allows us to create a much simpler pattern of
tests, and although we get some repetition, it's immediately clear for
each test what set of data is present when the test runs.

Finally, rework the ServiceFeedbackAggregator to raise an exception if
it's not going to run rather than return a string. The conditions
described are exceptional as they should not be triggered in normal
operation; it can only run twice per day if triggered manually, for
example (or if a cron job mistakenly runs twice) -- and it can only run
for today if it's triggered with a specific date or date range.