Exclude duplicate/non-actionable feedback from totals
Totals sent daily to the performance platform do not include those
pieces of ServiceFeedback that are marked as "non actionable" (or
duplicate).

This means that the aggregated feedback items shown to a user will not
match the data shown in the performance platform for a given period.

Therefore, we should exclude duplicate service feedback from the
aggregated service feedback as well.