Send page improvements to Zendesk
This receives feedback requests from the feedback frontend on behalf of the 'Improve this page' widget and forwards the feedback to Zendesk. https://trello.com/c/HN1L3EfK/286-update-footer-feedback-component

There isn't a requirement to browse the feedback anywhere apart from on Zendesk so we avoid the "Anonymous Feedback" flow in this app.

The 'Improve this page' feature is an experiment on the Service Manual. Traffic is expected to be very low. Therefore performance hasn't been taken seriously. Notably we are communicating with Zendesk synchronously. This will make the widget less snappy but has the advantage of being able to report to the user if it was successful.