Change how we deal with `params`
In rails 5 the params object is no longer a hash derivative.  This means
that in the places where we treated it as such we need to be explicit
about doing so by calling `.to_h` on it, otherwise we generate deprecation
warnings.  We also have to deal with behaviour that has changed where we
now get an empty string instead of a nil when asking for a param.