Declare content_item as optional on AnonymousContact
In rails 5 all `belongs_to` associations get a validation to check that
the target of the association actually exists.  You can disable this by
setting the `optional: true` option on the association.

It's not clear to me if `belongs_to :content_item` on `AnonymousContact`
is intented to be optional or not, but as it didn't already have a
validation in place, and tests were breaking because of the new validation
it seems fair to suggest it is supposed to be optional.