Provide rake task to fix misreported done page service feedback
For a couple of weeks in March 2017 the done page template sent the
service feedback form submissions to the support-api with the wrong
path - it used the slug of the service instead of the full path to
the done page (e.g. register-to-vote instead of /done/register-to-vote).
This meant that the service feedback was stored against the incorrect
url and couldn't be found in the feedback explorer for the done page.

Here we introduce an object that given a set of dates and a slug will
re-assign any service feedback (and aggregated service feedback) for
`/slug` to `/done/slug`.  There's also a version that will extract all
done pages that have some service feedback and fix all of them.

This feels safe to do because none of the /slug pages had a form on
them for submitting service feedback during the couple of weeks in
March so all service feedback in the db for /slug during that time
really is for /done/slug.  There are service feedback entries in the db
for some /slug pages outside those dates, but it's hard to know if those
are errors from similar bugs in the past, or from versions of the site
that did have service feedback forms on the /slug pages.  When we run
this script we'll limit it to 1st March 2017 to 1st April 2017.