Push fixed aggregated service feedback to perf platform
The system only pushes yesterday's aggregated service feedback to the
performance platform and doesn't look for changes it hasn't handled. This
means that after fixing the slugs in the aggregated feedback we have to
send it to the performance platform manually.  We do this inline rather
than run them as jobs to get immediate feedback.

We don't need to remove the old data from the performance platform
as it would never have been sent.  The code that gets the aggregated
details for sending them to the perf platform will only load data for
`/done/` paths and the misreported feedback didn't have `/done/` in
the path so wouldn't have been sent.