Stop raising when no aggregate feedback
When aggregate feedback for a time period is empty we were raising an
exception, seemingly to halt the flow and prevent the empty data being
sent to performance platform. There was a test that checked that the
error was swallowed when it was raised but not all callers of the code
were catching the error. This was causing it to be raised within the
worker and then reported multiple times to Sentry when the job was
retried.

This commit removes the exception and checks the `total` value from the payload
instead using that to decide whether to send to performance platform or
not.