Update README to bring it in line with styleguide
For: https://trello.com/c/O9BAVPvE/303-improve-readme-for-support-api-application

There's a styleguide for our [READMEs](https://docs.publishing.service.gov.uk/manual/readmes.html)
and the readme for this app was pretty out of date.  This is particularly annoying as the interplay
between the `support`, `support-api`, and `feedback` applications is confusing, so keeping the
readme uptodate is an opportunity to try to clarify things.