Adds multiple path search for Feedex
This will allow us to retrieve feedback for multiple paths in one go, instead of having to search for each path individually, exporting the files and then combining them together.

In order to search by multiple paths, a list of paths needs to be provided in the `path_prefixes` param. Example list of paths: path_prefixes: ['/vat-rates', '/guidance/', 'https://www.gov.uk/bank-holidays']

An email will be sent with the results of the search.

For: https://trello.com/c/IB9P5EyE/266-introduce-new-browse-options-for-feedex