Add and populate `document_type` field for ContentItem
The first migration will add the `document_type` field to ContentItem, which is meant to store the format of the pages where feedback was provided.

The second migration will populate the `document_type` field with values from the Content Store.

This will enable us to filter feedback by document type in Feedex.

For: https://trello.com/c/IB9P5EyE/266-introduce-new-browse-options-for-feedex