Mock passing of bearer token in controller tests
The controller tests don't automatically find and use the seeded user as
the request tests do.

`headers` isn’t a recognised key word in controller tests, so the
bearer token couldn’t be passed directly into to the request as they
could be in the requests tests.

Also, overwriting the value of `request.env[“Authorization”]` and
`request.headers[“Authorization”]` didn’t update the value used in
the controller tests.

The only way around the authentication is to stub the call to
`authenticate!` in gds-sso gem.

This follows the pattern established in Publishing API for mocking
authenticate:

* https://github.com/alphagov/publishing-api/blob/afdc1bf8b798e1b2f1754e1bf94a66a8c360d947/spec/rails_helper.rb#L65
* https://github.com/alphagov/publishing-api/blob/master/spec/support/authentication_helper.rb