Allow querying for anonymous feedback using POST
Currently the only way to perform a query against anonymous feeedback
using support-api is via a GET. This limits the number of paths that can
be passed in to the limit of a URL length. In practice, this means you
can only query on at most 30 URLs. The analytics team would like to be
able to run queries with more than 30 URLs.

This is the first part of a bigger to change which will also see the
support app use POST to run these queries.