Correctly allow POST requests to anonymous feedback endpoint
This follows on from 6035931bfc80bc3f274026d886cec38faa2a71c0.
Unfortunately that method appeared to work because the tests were
passing, but it doesn't actually set up the correct routes because of
the `only: :index`. There doesn't appear to be a way to allow the
`create` method of a resource to point to the same action as the `index`
action, so instead I've made the routes explicit.