Don't schedule Rake tasks outside of the Production environment
We're seeing a number of errors in Sentry as a result of at least
two workers (but probably more): `ProblemReportStatsPPUploaderWorker`
and `ProblemReportDailyTotalsPPUploaderWorker`.

These result in `GdsApi::SocketErrorException` timeouts on
Integration, most likely as a result of the nightly data sync,
where the dependent services (e.g. Performance Platform) are
temporarily down.

We shouldn't care about running these scheduled tasks on
Integration and Staging, particularly as they're all timed during
the data sync. It is better instead to let the tasks happen on
Production and wait for the nightly data sync to copy the results
over to Staging and Integration.

This does mean that Production may be 24 hours 'ahead' of the
other environments, but given that Performance Platform is being
retired, I don't see this as being an issue.

This should fix
https://sentry.io/organizations/govuk/issues/824216917/events/?project=202256&statsPeriod=24h
which has happened 20k times in 2 years.