Remove irrelevant section about import orgs
This won't work locally as the organisations API relies on special
public-facing routes [1]. Since the instructions are broken and also
unclear about exactly when it makes sense to do a manual import, I
think it makes sense to remove them pending a clear need. In any case,
a developer can still populate the DB manually or via replication.

[1]: https://github.com/alphagov/gds-api-adapters/blob/5a0e77887ca1c959f11bc8cd9690b62835707a54/lib/gds_api.rb#L141