Allow any app to create support tickets
We want to make it easy for any app to create a support ticket
and support-api feels like the best place for this.

The endpoint `/support-tickets` will be made available to apps via
`gds-api-adapters`, which allows them to pass Zendesk ticket
parameters to support-api.

The `gds_zendesk` gem provides a client to communicate with Zendesk
and it has a `ticket.create!` method, which is called if the parameters
provided are valid.

The only required standard ticket fields are `subject` and `description`
- for more information see https://support.zendesk.com/hc/en-us/articles/4408886739098-About-ticket-fields

This provides the basis of creating Zendesk tickets and will need
adjustments once we know what other information is required for
different types of support tickets.

We should keep the naming convention generic, as there is a high
possibility that we move away from Zendesk in the future.