Use the comment property to set the Zendesk ticket description
This was causing an error
   GdsApi::HTTPUnprocessableEntity: URL: http://support-api/support-tickets
   Response body:
   {"status":"error","errors":{"description":["can't be blank"]}}

The docs say to use `comment` instead.

https://developer.zendesk.com/api-reference/ticketing/tickets/tickets/#description-and-first-comment

"When creating a ticket, use the comment property to set the ticket description, which is also the first comment.
Important: Do not use the description property to set the first comment. The property is for reading purposes only. While it's possible to use the property to set the first comment, the functionality has limitations and is provided to support existing implementations."

Ticket Comments are represented as JSON objects. They have a multiple properties.
The comment string needs to be in the "body" property.
See https://developer.zendesk.com/api-reference/ticketing/tickets/ticket_comments/