Don't format ticket body in SupportTicket
We anticipate that the inbound API call will have the body formatted so we don't need to do it here.

`user_agent` attribute should also be removed from support_tickets_controller.
Since it requires update to the Pact test in a another repo it will be removed in a follow up PR.