Check if requester is suspended before creating Zendesk ticket
It's not possible to create Zendesk ticket if a requester is suspended in
Zendesk. The following error is raised:
  ZendeskAPI::Error::RecordInvalid:
  {"requester"=>[{"description"=>"Requester: <Name> is suspended."}]}
  (ZendeskAPI::Error::RecordInvalid)

We don't want to rescue ZendeskAPI::Error::RecordInvalid as this error is
raised for other issue such as incorrect formatting.

Suspended users can be found https://govuk.zendesk.com/agent/user_filters