Use ENV vars instead of deprecated `Rails.application.secrets`
`Rails.application.secrets` was deprecated and was removed in rails 7.2
https://github.com/rails/rails/pull/47801

It is safe to delete `secrets.yml` as Rails first reads the value from
SECRET_KEY_BASE environment variable which is set in govuk-helm-charts (and
then from credentials file).
https://github.com/rails/rails/blob/v7.2.1/railties/lib/rails/application.rb#L453-L471

In development and test the value is randomly generated and stored in a
temporary file in `tmp/local_secret.txt`.

The `ENV SECRET_KEY_BASE_DUMMY=1` is set in
https://github.com/alphagov/govuk-ruby-images/blob/main/builder.Dockerfile#L16