Do not report team only requests to Sentry in integration and staging
We have set up our Notify accounts to not send emails in integration and
staging, unless the email address is a considered a “team member”.  This
is deliberate to stop users from receiving duplicated emails from
production plus integration/staging.

Notify returns an error if you try to send an email address is someone
who isn’t a team member:

```
Notifications::Client::BadRequestError
BadRequestError: Can't send to this recipient using a team-only API key
```

In most apps (e.g. in Whitehall and Publisher) we don’t log these to
Sentry, as they serve no purpose and are not actionable.

Making the same change here, so these are no longer logged to Sentry in
those environments.