Optimise and simplify anonymous contact summaries
Summaries by organisation for an organisation with a lot of feedback
required ~600-1400ms time spent querying a warm local database, and much
longer in a cold one. In production, if the database hadn't recently
been queried for the organisation's feedback, Support API response time
could be as slow as 20 seconds, by which time the Support app's request
(proxied via gds-api-adapters) would have long since timed out. A second
request to the page after Support API had finished its query would tend
to work for some time, suggesting some success with caching but not a
great overall user experience

I tried EXPLAINing and even EXPLAIN ANALYZEing the query in ContentItem,
but nothing jumped out at me as an obvious optimisation (I'm not an
expert)

Instead, I had a go at rewriting this in a simpler way and starting with
the type of object we were really interested in - AnonymousContact -
rather than ContentItem. This has brought local loads down to 150-200ms
for the previously ill-performing page. I can't pretend to know exactly
why this approach is faster, but maybe the previous complexity was
making it hard for Postgres' query planner to identify an optimum
strategy - an EXPLAIN did identify a bad estimate on a costly step

Perhaps another optimisation lies in removing the ordering and group
filter from the ActiveRecord query in favour of sorting and filtering
the smaller and simpler array of count hashes outside of the database

In opting for a less 'clever' approach, we're now issuing one query per
time range. My local server now reports 8 queries with 3 cached, which
is a small increase from 6 and 3. But either way, speed is speed...

... and I think this is easier to read?

One final difference: in the previous aproach the 'to' time was 00:00
today, whereas in the new approach the 'to' time is 23:59:59.999999999
yesterday. This is a trade off of using the existing
AnonymousContact.created_between_days scope method, and one with which I
think we can live