Remove impossible case test
I don't think this is a possible case. This is getting the document type
from a URL param. Without the param the path would hit the index action.
As far as I know there's no way to hit this action with a nil document
type

Note that the test description looks like it was a copy/paste from an
equivalent organisations controller test