Fix `ArgumentError`
Whilst upgrading to Ruby 3.4, using `Hash.new(args...)` resulted in an
error: `ArgumentError: unknown keywords: :path_prefixes, :from, :to`.

It appears this usage was undocumented. As the intention was unclear, I
assume the original author was trying to create a hash with these
values.

Therefore replacing `Hash.new` with standard hash syntax.